compensationInsights/requirements.txt
freeplay>=0.5.0
anthropic
openai
python-dotenv
